4.1.13 remove utility functions from task-mock
