+++ src/utils/utils.js
@@ -32,4 +32,24 @@ const render = (container, element, place = RenderPosition.BEFOR_EEND) => {
   }
 };
 
-export {formatTime, createElement, render, RenderPosition};
+const getRandomSign = () => Math.random() > 0.5 ? 1 : -1;
+
+const getRandomIntegerNumber = (min, max) => {
+  return min + Math.floor(Math.random() * (max - min));
+};
+
+const getRandomArrayItems = (array) => {
+  const randomIndex = getRandomIntegerNumber(0, array.length);
+
+  return array[randomIndex];
+};
+
+const getRandomDate = () => {
+  const targetDate = new Date();
+  const diffValue = getRandomSign() * getRandomIntegerNumber(0, 8);
+  targetDate.setDate(targetDate.getDate() + diffValue);
+
+  return targetDate;
+};
+
+export {formatTime, createElement, render, RenderPosition, getRandomIntegerNumber, getRandomArrayItems, getRandomDate};
